Refactor shipping label code to use child containers
