+++ client/apps/shipping-label/style.scss
@@ -12,11 +12,11 @@
 		margin-top: 8px;
 	}
 
-	.shipping-label__item {
+	.shipping-label__item, .label-item {
 		padding: 16px;
 		border-bottom: 1px solid #eee;
 
-		.shipping-label__item-created {
+		.label-item__created {
 			color: $gray;
 			font-size: 12px;
 		}
@@ -29,7 +29,7 @@
 			margin-bottom: 3px;
 			text-align: left;
 
-			&.shipping-label__item-tracking {
+			&.label-item__tracking {
 				margin-bottom: 16px;
 				font-size: 12px;
 
@@ -39,7 +39,7 @@
 				}
 			}
 
-			& .shipping-label__item-detail {
+			& .label-item__detail {
 				font-weight: bold;
 				font-size: 14px;
 				color: $gray-dark;
@@ -55,7 +55,7 @@
 			margin-top: 16px;
 		}
 
-		.shipping-label__item-actions {
+		.label-item__actions {
 			display: flex;
 			justify-content: space-between;
 			font-size: 12px;
